viewcache template + styles cleanup
