+++ lib/common_tpl_funcs.php
@@ -1,6 +1,7 @@
 <?php
 
 use Utils\View\View;
+use Utils\Uri\Uri;
 
 //set the global template-name variable
 function tpl_set_tplname($local_tpl_name){
@@ -150,6 +151,7 @@ function tpl_BuildTemplate($dbdisconnect = true, $minitpl = false, $noCommonTemp
     global $stylepath, $tplname, $vars, $langpath, $lang_array, $lang, $language, $menu, $config, $usr;
 
     // object
+    /** @var View $view */
     global $view;
 
     //language specific expression
@@ -195,6 +197,10 @@ function tpl_BuildTemplate($dbdisconnect = true, $minitpl = false, $noCommonTemp
         $sCode = file_get_contents($stylepath . '/main.tpl.php');
     }
 
+    //global css files:
+    $view->setVar('screenCss', Uri::getLinkWithModificationTime('tpl/stdstyle/css/style_screen.css'));
+    $view->setVar('printCss', Uri::getLinkWithModificationTime('tpl/stdstyle/css/style_print.css'));
+    $view->setVar('seasonCss', Uri::getLinkWithModificationTime('tpl/stdstyle/css/style_'.$view->getSeasonCssName().'.css'));
 
     //does template exist?
     if (!file_exists($stylepath . '/' . $tplname . '.tpl.php')) {
