+++ lib/languages/pl.php
@@ -923,7 +923,7 @@ $translations = array(
     'openchecker_degrees' => 'Stopnie',
     'openchecker_minutes' => 'minuty',
     'openchecker_enable_checkbox' => 'Zaznaczenie tego elementu umożliwi użytkownikom sprawdzanie otrzymanych współrzędnych przy pomocy OpenSprawdzacza',
-    'openchecker_enabled' => 'Autor tego geocache umożliwił sprawdzenie poprawności rozwiązania zagadki (współrzędnych pojemnika finałowego) przed wyruszeniem w teren!',
+    'openchecker_enabled' => 'Autor tej skrytki umożliwił sprawdzenie poprawności rozwiązania zagadki (współrzędnych pojemnika finałowego).',
     'openchecker_tries' => 'prób',
     'openchecker_hits' => 'trafień',
     'openchecker_attempts_01' => 'próbowałeś zgadywać',
