+++ tpl/stdstyle/css/style_print.css
@@ -2,7 +2,7 @@ body, html {
     background-color: white;
 }
 
-.noprint, .navflag-container, .site-slogan-container, .nav1-container, .footer, .page-container-1-2, #logo, #bg1, #bg2, .header, .site-name, .nav2, .nav3, .notlogged-cacheview, #cachemenu, #viewcache-attributesend
+.noprint, .navflag-container, .site-slogan-container, .nav1-container, .footer, .page-container-1-2, #logo, #bg1, #bg2, .header, .site-name, .nav2, .nav3, .notlogged-cacheview, #cachemenu
 {
     display: none;
     width:0;
@@ -21,9 +21,6 @@ body, html {
 .content2-container, .content3-container {clear: both; float: none; width: 769px; margin: 0px 0px 15px 0px; padding: 0px;}
 .line-box {width:769px; border-bottom: solid 1px rgb(200,200,200); padding-bottom: 7px; margin: 0px 0px 15px 0px;}
 
-div.viewcache-pictureblock { float: left; }
-
-div.viewcache-pictureblock span.title { display: block; max-width: 175px; padding-left: 10px; padding-right: 5px;clear: left; font-size: 12px;}
 div.frame {
     width: 94%;
     margin: 5px;
