+++ tpl/stdstyle/main.tpl.php
@@ -48,9 +48,10 @@ if (date('m') == 12 || date('m') == 1) {
         <meta name="keywords" content="geocaching, opencaching, skarby, poszukiwania, geocashing, longitude, latitude, utm, coordinates, treasure hunting, treasure, GPS, global positioning system, garmin, magellan, mapping, geo, hiking, outdoors, sport, hunt, stash, cache, geocaching, geocache, cache, treasure, hunting, satellite, navigation, tracking, bugs, travel bugs">
         <meta name="author" content="{site_name}">
 
-        <link rel="stylesheet" type="text/css" media="screen" href="tpl/stdstyle/css/style_screen.css">
-        <link rel="stylesheet" type="text/css" media="print" href="tpl/stdstyle/css/style_print.css">
-        <link rel="stylesheet" type="text/css" media="screen" href="tpl/stdstyle/css/style_{season}.css">
+        <link rel="stylesheet" type="text/css" media="screen" href="<?=$view->screenCss?>">
+        <link rel="stylesheet" type="text/css" media="screen" href="<?=$view->seasonCss?>">
+        <link rel="stylesheet" type="text/css" media="print" href="<?=$view->printCss?>">
+
 
         <link rel="shortcut icon" href="/images/<?=$config['headerFavicon']?>">
         <link rel="apple-touch-icon-precomposed" href="/images/oc_logo_144.png">
