+++ tpl/stdstyle/mini.tpl.php
@@ -12,9 +12,8 @@
         <meta http-equiv="cache-control" content="no-cache" />
         <meta name="KEYWORDS" content="geocaching, opencaching, skarby, poszukiwania, geocashing, longitude, latitude, utm, coordinates, treasure hunting, treasure, GPS, global positioning system, garmin, magellan, mapping, geo, hiking, outdoors, sport, hunt, stash, cache, geocaching, geocache, cache, treasure, hunting, satellite, navigation, tracking, bugs, travel bugs" />
         <meta name="author" content="Opencaching.pl " />
-        <meta http-equiv="X-UA-Compatible" content="IE=10" />
-        <link rel="stylesheet" type="text/css" media="screen,projection" href="tpl/stdstyle/css/style_screen.css" />
-        <link rel="stylesheet" type="text/css" media="print" href="tpl/stdstyle/css/style_print.css" />
+        <link rel="stylesheet" type="text/css" media="screen,projection" href="<?=$view->screenCss?>" />
+        <link rel="stylesheet" type="text/css" media="print" href="<?=$view->printCss?>" />
         <link rel="SHORTCUT ICON" href="favicon.ico" />
         <link rel="apple-touch-icon-precomposed" href="/images/oc_logo_144.png" />
